Updated documentation to reflect refactor
Updates the documentation to reflect the elimination of custom response
types, and the fact that the factory now accepts the `$request`
instance, and not the `Accept` header.

Also adds a quick start, and re-organizes the documentation to reflect
the three distinct features: response generation, custom exceptions, and
error handling middleware.