Refactors ProblemDetailsResponseFactory to use a response factory
Instead of composing a response prototype and a stream factory, it now
composes a response _factory_, which **MUST** be supplied during
instantiation.

This has other implications: the various classes that use a
`ProblemDetailsResponseFactory` now can no longer create them without
arguments. As such, such constructor arguments are no longer optional,
but required.

The side benefit is that this completely de-couples the library from
Diactoros!